refactor(status): move fake-data.ts to src/dev/fake-status.ts
Make the runtime/dev intent of the fixture explicit in its path.
The Worker still imports fakeStatus() at runtime when
USE_FAKE_DATA=true; only the location and filename change.
